hook 12 pasos creado
